Move tests to lsk.test.js
